Fork awesome magnifier layout
The awesome version of this layout rearanges the windows when the focus
changes. I do not like that. My own version keeps the second window
always in the middle (this means that it gets hidden when the main
window gets focus again but I usually just use the layout to do some quick
tasks in a terminal so that doesn't matter)